Soft-close Gain-node Issue
Implementation of the rangeHandler() has seemingly fixed the gain node
issue. Since full diagnosis of the cause of the issue was never found,
now that it has disappeared it is difficult to say if it is fully fixed.

Testing results in no issues during fast chord changes, which is the
sole place the error occured.